repo: move puzzle models into folder
